refactor: create hook and update pages
